+++ documentation/changelog.markdown
@@ -14,6 +14,7 @@ To be determined.
 - **New**: Added legend delegate methods to customize the swatch fill and border line style for each legend entry.
 - **New**: Added a plot property to control drawing of the legend entry swatch.
 - **New**: Added options to draw a background fill behind and a border line around each legend entry.
+- **New**: Added an axis property to control the offset direction of the title independent of the tick direction.
 - **Changed**: Switched to a new algorithm for computing curved scatter plots.
 - **Changed**: Miscellaneous bug fixes and cleanup.
 - **Removed**: Removed the elastic global range properties from `CPTXYPlotSpace`. Turning on momentum scrolling now automatically allows elastic ranges, too.